Rename variations declaration controller to package playback.DeclarationController
